posts updated

+++ client/src/actions/postAction.js
@@ -1,5 +1,6 @@
 import axios from '../Config/axios'
 import { setAlert } from './alertAction'
+import Swal from 'sweetalert2'
 
 export const getAllPosts = (post) => {
 	return { type: 'GET_POSTS', payload: post }
@@ -7,6 +8,7 @@ export const getAllPosts = (post) => {
 export const getsinglePost = (post) => {
 	return { type: 'GET_POST', payload: post }
 }
+export const UPDATE_LIKES = 'UPDATE_LIKES'
 
 export const upDateLikes = (id, likes) => {
 	return { type: 'UPDATE_LIKES', payload: { id, likes } }
@@ -39,10 +41,8 @@ export const getPosts = () => {
 					dispatch(getAllPosts(post))
 				})
 				.catch((err) => {
-					const errors = err.response.data.errors
-
-					if (errors) {
-						errors.forEach((error) => dispatch(setAlert(error.msg, 'danger')))
+					if (err) {
+						dispatch(setAlert('err', 'danger'))
 					}
 				})
 		} catch (err) {
@@ -77,20 +77,34 @@ export const startGetPost = (id) => {
 	}
 }
 
-//aDD LIKE
+// //aDD LIKE
 export const addLike = (id) => {
 	return (dispatch) => {
 		try {
 			axios
-				.put(`/posts/likes/${id}`, {
-					headers: {
-						Authorization: localStorage.getItem('authToken'),
-					},
-				})
+				.put(
+					`/posts/likes/${id}`,
+					{},
+					{
+						headers: {
+							Authorization: localStorage.getItem('authToken'),
+						},
+					}
+				)
 				.then((response) => {
 					const likes = response.data
 					//console.log(likes)
-					dispatch(upDateLikes())
+					dispatch(upDateLikes(id, likes))
+					if (response.data.msg) {
+						Swal.fire({
+							text: 'Post Has Already been liked!',
+							showConfirmButton: false,
+							icon: 'warning',
+							toast: true,
+							timer: 2000,
+							position: 'top',
+						})
+					}
 				})
 				.catch((err) => {
 					const errors = err.response.data.errors
@@ -107,21 +121,32 @@ export const removeLike = (id) => {
 	return (dispatch) => {
 		try {
 			axios
-				.put(`/posts/unlikes/${id}`, {
-					headers: {
-						Authorization: localStorage.getItem('authToken'),
-					},
-				})
+				.put(
+					`/posts/unlikes/${id}`,
+					{},
+					{
+						headers: {
+							Authorization: localStorage.getItem('authToken'),
+						},
+					}
+				)
 				.then((response) => {
+					//console.log(response.data)
 					const likes = response.data
 					dispatch(upDateLikes(id, likes))
-				})
-				.catch((err) => {
-					const errors = err.response.data.errors
-
-					if (errors) {
-						errors.forEach((error) => dispatch(setAlert(error.msg, 'danger')))
+					if (response.data.msg) {
+						Swal.fire({
+							text: 'Post Has Not yet been liked!',
+							showConfirmButton: false,
+							icon: 'warning',
+							toast: true,
+							timer: 2000,
+							position: 'top',
+						})
 					}
+				})
+				.catch((res) => {
+					console.log(res.data)
 				})
 		} catch (err) {
 			console.log(err)
@@ -133,23 +158,46 @@ export const removeLike = (id) => {
 export const startDeletePosts = (id) => {
 	return (dispatch) => {
 		try {
-			axios
-				.delete(`posts/${id}`, {
-					headers: {
-						Authorization: localStorage.getItem('authToken'),
-					},
-				})
-				.then(() => {
-					dispatch(deletePost(id))
-					dispatch(setAlert('Post Removed', 'success'))
-				})
-				.catch((err) => {
-					const errors = err.response.data.errors
+			Swal.fire({
+				title: 'Are Your Sure?',
+				icon: 'warning',
+				text: "You won't be able to revert this!",
+				showDenyButton: true,
+				focusConfirm: false,
+				confirmButtonText: '<i class="fa fa-thumbs-up"></i> Yes, delete it!',
+				confirmButtonAriaLabel: 'Thumbs up, great!',
+				denyButtonText: '<i class="fa fa-thumbs-down"> Cancel </i>',
+				denyButtonAriaLabel: 'Thumbs down',
+			}).then((result) => {
+				if (result.isConfirmed) {
+					axios
+						.delete(`posts/${id}`, {
+							headers: {
+								Authorization: localStorage.getItem('authToken'),
+							},
+						})
+						.then(() => {
+							dispatch(deletePost(id))
+							Swal.fire({
+								text: 'Post Removed Successfully!',
+								showConfirmButton: false,
+								icon: 'error',
+								toast: true,
+								timer: 2000,
+								position: 'top',
+							})
+						})
+						.catch((err) => {
+							const errors = err.response.data.errors
 
-					if (errors) {
-						errors.forEach((error) => dispatch(setAlert(error.msg, 'danger')))
-					}
-				})
+							if (errors) {
+								errors.forEach((error) =>
+									dispatch(setAlert(error.msg, 'danger'))
+								)
+							}
+						})
+				}
+			})
 		} catch (err) {
 			console.log(err)
 		}
@@ -169,7 +217,14 @@ export const startAddPost = (form) => {
 				.then((response) => {
 					const post = response.data
 					dispatch(AddPost(post))
-					dispatch(setAlert('Post Created', 'success'))
+					Swal.fire({
+						text: 'Post Added Successfully',
+						showConfirmButton: false,
+						icon: 'success',
+						toast: true,
+						timer: 2000,
+						position: 'top',
+					})
 				})
 				.catch((err) => {
 					console.log(err)
@@ -194,7 +249,14 @@ export const startAddComment = (postId, formdata) => {
 				.then((response) => {
 					const comment = response.data
 					dispatch(AddComment(comment))
-					dispatch(setAlert('Comment Added', 'success'))
+					Swal.fire({
+						text: 'Comment Added',
+						showConfirmButton: false,
+						icon: 'success',
+						toast: true,
+						timer: 2000,
+						position: 'top',
+					})
 				})
 				.catch((err) => {
 					const errors = err.response.data.errors
@@ -212,29 +274,44 @@ export const startAddComment = (postId, formdata) => {
 export const startRemoveComment = (postId, commentId) => {
 	return (dispatch) => {
 		try {
-			axios
-				.delete(`/posts/comment/${postId}/${commentId}`, {
-					headers: {
-						'Content-type': 'application/json',
-						Authorization: localStorage.getItem('authToken'),
-					},
-				})
-				.then((response) => {
-					if (response.data.hasOwnProperty('errors')) {
-						const errors = response.data.errors
-						//console.log(errors)
-						if (errors) {
-							dispatch(setAlert(errors, 'danger'))
-						} else {
-							//const ceomment = response.data
+			Swal.fire({
+				title: 'Are Your Sure?',
+				icon: 'warning',
+				text: "You won't be able to revert this!",
+				showDenyButton: true,
+
+				focusConfirm: false,
+				confirmButtonText: '<i class="fa fa-thumbs-up"></i> Yes ,delete it',
+				confirmButtonAriaLabel: 'Thumbs up, great!',
+				denyButtonText: '<i class="fa fa-thumbs-down"> Cancel</i>',
+				denyButtonAriaLabel: 'Thumbs down',
+			}).then((result) => {
+				if (result.isConfirmed) {
+					axios
+						.delete(`/posts/comment/${postId}/${commentId}`, {
+							headers: {
+								'Content-type': 'application/json',
+								Authorization: localStorage.getItem('authToken'),
+							},
+						})
+						.then((response) => {
+							//const comment = response.data
 							dispatch(RemoveComment(commentId))
-							dispatch(setAlert('Comment Removed', 'success'))
-						}
-					}
-				})
-				.catch((err) => {
-					console.log(err)
-				})
+							console.log('comment removed')
+							Swal.fire({
+								text: 'Comment Removed',
+								showConfirmButton: false,
+								icon: 'success',
+								toast: true,
+								timer: 2000,
+								position: 'top',
+							})
+						})
+						.catch((err) => {
+							console.log(err)
+						})
+				}
+			})
 		} catch (err) {
 			console.log(err)
 		}
